cloudfoundry-worker/src/main/java/io/pivotal/conductor/worker/cloudfoundry/CreateCloudFoundrySpaceWorker.java
package io.pivotal.conductor.worker.cloudfoundry;

import com.netflix.conductor.client.worker.Worker;
import com.netflix.conductor.common.metadata.tasks.Task;
import com.netflix.conductor.common.metadata.tasks.TaskResult;
import com.netflix.conductor.common.metadata.tasks.TaskResult.Status;

public class CreateCloudFoundrySpaceWorker implements Worker {

    static final String TASK_DEF_NAME = "create_cloud_foundry_space";

    @Override
    public String getTaskDefName() {
        return TASK_DEF_NAME;
    }

    @Override
    public TaskResult execute(Task task) {

        String projectName = (String) task.getInputData().get("projectName");
        String spaceNameSuffix = (String) task.getInputData().get("spaceNameSuffix");
        String spaceName = CloudFoundryUtil.deriveSpaceName(projectName, spaceNameSuffix);

        // create space on cloud foundry

        TaskResult taskResult = new TaskResult(task);
        taskResult.setStatus(Status.COMPLETED);
        taskResult.getOutputData().put("spaceName", spaceName);

        return taskResult;
    }
}
